Add a img, a paragraph, unordered lis, and another paragraph
The image addedwas a picture fo the recipe itself, amcaroni and cheese. I copied the image address and directly
put it into the src attribute. The paragraph contained just a brief description of the recipe and the benefit
recieved if the reader were to make it. Next, I created an unordered list for the ingredients, where each
one of the list items (li) contained a single ingredient. Finally, the last paragraph I created contained steps
that serve as instructions on how to make the recipe.