Add more recipe pages and link to beef stew, pizza
I used the same page format to create two more recipes. This included adding an image toeach of the recipe pages, as well
as adding a description, ingredients, and recipe steps for each. I also had to add a link to each of these new pages on
the main index.html page.